+++ packages/sdk/src/index.ts
@@ -228,6 +228,7 @@ export {
   finalizeUnwrapContract,
   setOperatorContract,
   underlyingContract,
+  inferredTotalSupplyContract,
   wrapContract,
   supportsInterfaceContract,
   isConfidentialTokenContract,